fix: Make `davRootPath` and `davRemoteURL` support public shares
