+++ package.json
@@ -67,11 +67,11 @@
   },
   "dependencies": {
     "@nextcloud/auth": "^2.3.0",
-    "@nextcloud/initial-state": "^2.2.0",
     "@nextcloud/l10n": "^3.1.0",
     "@nextcloud/logger": "^3.0.2",
     "@nextcloud/paths": "^2.1.0",
     "@nextcloud/router": "^3.0.1",
+    "@nextcloud/sharing": "^0.2.1",
     "cancelable-promise": "^4.3.1",
     "is-svg": "^5.0.1",
     "typescript-event-target": "^1.1.1",
